requirements.txt
langgraph
langchain-google-genai
langchain-core
langchain-community
python-dotenv
google-ai-generativelanguage>=0.6.17
faiss-cpu